Support response.done and add integration tests
